IndexedDB/tests/submissions/Opera/idbcursor-direction.htm

<!DOCTYPE html>
<title>IDBCursor.direction</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>

    function cursor_direction(dir)
    {
        var db,
          t = async_test(document.title + " - " + dir),
          expected = dir ? dir : IDBCursor.NEXT;

        var open_rq = createdb(t);
        open_rq.onupgradeneeded = function(e) {
            db = e.target.result;
            var objStore = db.createObjectStore("test");

            objStore.add("data", "key");
        };

        open_rq.onsuccess = t.step_func(function(e) {
            var cursor_rq = db.transaction("test")
                              .objectStore("test")
                              .openCursor(undefined, dir);

            cursor_rq.onsuccess = t.step_func(function(e) {
                var cursor = e.target.result;

                assert_equals(cursor.direction, expected, 'direction');
                assert_readonly(cursor, 'direction')

                t.done();
            });
        });
    }

    cursor_direction();
    cursor_direction(IDBCursor.NEXT);
    cursor_direction(IDBCursor.NEXT_NO_DUPLICATE);
    cursor_direction(IDBCursor.PREV);
    cursor_direction(IDBCursor.PREV_NO_DUPLICATE);

</script>

<div id="log"></div>
